gitlab-ci.yml: Run test with KVM and macOS via cirrus-run
Since Travis changed their policy, we soon cannot use it for the
kvm-unit-tests anymore, thus we lose the CI testing with KVM enabled
and the compilation jobs on macOS. Fortunately there is an alternative:
Cirrus-CI also provides containers with KVM enabled and CI jobs with
macOS. Thanks to the so-call "cirrus-run" script, we can even start
the jobs from the gitlab-CI, so we get all the test coverage in the
gitlab-CI again. cirrus-run needs some configuration first, though.
Please refer to the description from libvirt for the details how to
set up your environment for these jobs:

 https://gitlab.com/libvirt/libvirt/-/blob/v7.0.0/ci/README.rst

Signed-off-by: Thomas Huth <thuth@redhat.com>
Message-Id: <20210115223017.271339-1-thuth@redhat.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>